Add Cudd_bddNot function.
Cudd_Not is a macro and therefore not available if CUDD is compiled
as a dynamic library.

This commit introduces Cudd_bddNot that complements the current BDD
API in a way similar to bddAnd and bddOr. That is Cudd_bddNot takes
the manager as first argument, even if not used in practice.